main-compose/requirements.txt
flask
mysqlclient 
mysql-connector-python-rf
mysql-connector-python
